fix msgId generator

+++ util/MsgId-gen_test.go
@@ -8,9 +8,9 @@ import (
 func TestMsgIdGen(t *testing.T) {
 	var (
 		UID          string = "haha"
-		registerTime int = 5
-		lastMsgTime  int = 10
-		tip          int = 0
+		registerTime int64 = 5
+		lastMsgTime  int64 = 10
+		tip          int64 = 0
 		res 		 string = ""
 	)
 	lastMsgTime,tip,res = MsgIdGen(UID, registerTime, lastMsgTime, tip)
